pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-medical-guard"
version = "0.1.0"
description = "Guardrails for LLM-generated medical and health content"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Pillright", email = "dev@pillright.com" }]
keywords = [
    "llm",
    "guardrails",
    "medical",
    "health",
    "safety",
    "ai",
    "content-moderation",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
langchain = ["langchain-core>=0.2"]
openai = ["openai>=1.0"]
all = ["langchain-core>=0.2", "openai>=1.0"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/pillright/llm-medical-guard"
Documentation = "https://github.com/pillright/llm-medical-guard#readme"
Repository = "https://github.com/pillright/llm-medical-guard"
Issues = "https://github.com/pillright/llm-medical-guard/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_medical_guard"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
strict = true
